put every component inside folder to make it clean
