+++ services-platform-frontend/src/Components/categories/categories.css
@@ -1,33 +1,3 @@
-.home-container{
-    display: grid;
-    grid-template-columns: 2fr 10fr;
-}
-
-.main-hero{
-    display: flex;
-    gap: 20px;
-    font-weight: 600;
-    padding: 5px;
-    font-family: 'Fira Sans', sans-serif;
-    background-repeat: no-repeat;
-    background-image: src="./images/hero-img.jpg";
-    height: 40%;
-    width: 100%;
-    background-size: 100%;
-}
-
-.hero-text{
-    margin-top: 5%;
-    margin-left: 5%;
-}
-
-.hero-text h1{
-    font-size: 3rem;
-}
-
-.hero-text p{
-    font-size: 1.3rem;
-}
 
 .category-container{
     margin-top: 30px;
@@ -75,7 +45,6 @@
     gap: 10px;
     justify-content: center;
 }
-
 
 .icon-txt{
     margin-left: 10px;
