+++ services-platform-frontend/src/Components/sidebar/sidebar.css
